move flaghelper package, add new constructor, update comments
