+++ pflag/pflag_test.go
@@ -340,7 +340,7 @@ func TestPFlags(t *testing.T) {
 	} {
 		tbl := itbl
 		t.Run(tbl.name, func(t *testing.T) {
-			// t.Parallel()
+			t.Parallel()
 			ctx := context.Background()
 			// use UpperSnakeCase instead of default (CamelCase) since
 			// single character field names like A and B make it hard to decode
